requirements.txt
torch
torchaudio
pytorch-lightning
librosa
soundfile
openai-whisper
tqdm
matplotlib
numpy
cython